requirements.txt
fastapi
uvicorn
pillow
python-multipart
tensorflow-cpu
numpy
